+++ src/Symfony/Component/Translation/PluralizationRules.php
@@ -15,6 +15,8 @@ namespace Symfony\Component\Translation;
  * Returns the plural rules for a given locale.
  *
  * @author Fabien Potencier <fabien@symfony.com>
+ *
+ * @deprecated since Symfony 4.2, use IdentityTranslator instead
  */
 class PluralizationRules
 {
@@ -28,8 +30,12 @@ class PluralizationRules
      *
      * @return int The plural position
      */
-    public static function get($number, $locale)
+    public static function get($number, $locale/*, bool $triggerDeprecation = true*/)
     {
+        if (3 > \func_num_args() || \func_get_arg(2)) {
+            @trigger_error(sprintf('The "%s" class is deprecated since Symfony 4.2.', __CLASS__), E_USER_DEPRECATED);
+        }
+
         if ('pt_BR' === $locale) {
             // temporary set a locale for brazilian
             $locale = 'xbr';
@@ -196,6 +202,8 @@ class PluralizationRules
      */
     public static function set(callable $rule, $locale)
     {
+        @trigger_error(sprintf('The "%s" class is deprecated since Symfony 4.2.', __CLASS__), E_USER_DEPRECATED);
+
         if ('pt_BR' === $locale) {
             // temporary set a locale for brazilian
             $locale = 'xbr';