+++ src/Symfony/Component/Translation/Tests/MessageSelectorTest.php
@@ -14,6 +14,9 @@ namespace Symfony\Component\Translation\Tests;
 use PHPUnit\Framework\TestCase;
 use Symfony\Component\Translation\MessageSelector;
 
+/**
+ * @group legacy
+ */
 class MessageSelectorTest extends TestCase
 {
     /**