+++ src/Symfony/Contracts/CHANGELOG.md
@@ -5,3 +5,4 @@ CHANGELOG
 -----
 
  * added `Service\ResetInterface` to provide a way to reset an object to its initial state
+ * added `Translation\TranslatorInterface` and `Translation\TranslatorTrait`